requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0
faiss-cpu>=1.7.4
tqdm>=4.65.0
flask>=2.3.0
flask-cors>=4.0.0